move around some files and add in middleware support
