+++ src/background/NoIntrospection.ts
